Move upload helpers into dedicated module
